+++ CopilotKit/packages/react-core/src/types/coagent-action.ts
@@ -13,5 +13,7 @@ export type CoagentAction<T = any> = {
   name: string;
   nodeName?: string;
   handler?: (props: CoagentActionHandlerArguments<T>) => void | Promise<void>;
-  render?: ((props: CoagentActionRenderProps<T>) => string | React.ReactElement) | string;
+  render?:
+    | ((props: CoagentActionRenderProps<T>) => string | React.ReactElement | undefined | null)
+    | string;
 };