+++ examples/coagents-research-canvas/ui/src/components/Progress.tsx
@@ -1,0 +1,53 @@
+import { cn } from "@/lib/utils";
+import { CheckIcon, LoaderCircle } from "lucide-react";
+
+const truncateUrl = (url: string, maxLength: number = 40) => {
+  if (url.length <= maxLength) return url;
+  return url.substring(0, maxLength - 3) + "...";
+};
+
+export function Progress({
+  logs,
+}: {
+  logs: {
+    message: string;
+    done: boolean;
+  }[];
+}) {
+  if (logs.length === 0) {
+    return null;
+  }
+
+  return (
+    <div>
+      <div className="border border-slate-200 bg-slate-100/30 shadow-md rounded-lg overflow-hidden text-sm py-2">
+        {logs.map((log, index) => (
+          <div key={index} className="flex">
+            <div className="w-8">
+              <div className="w-4 h-4 bg-slate-700 flex items-center justify-center rounded-full mt-[10px] ml-[12px]">
+                {log.done ? (
+                  <CheckIcon className="w-3 h-3 text-white" />
+                ) : (
+                  <LoaderCircle className="w-3 h-3 text-white animate-spin" />
+                )}
+              </div>
+              {index < logs.length - 1 && (
+                <div
+                  className={cn("h-full w-[1px] bg-slate-200 ml-[20px]")}
+                ></div>
+              )}
+            </div>
+            <div className="flex-1 flex justify-center py-2 pl-2 pr-4">
+              <div className="flex-1 flex items-center text-xs">
+                {log.message.replace(
+                  /https?:\/\/[^\s]+/g, // Regex to match URLs
+                  (url) => truncateUrl(url) // Replace with truncated URL
+                )}
+              </div>
+            </div>
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+}